Переименовывает файлы, удаляет расширения файлов из импортов
