+++ src/components/loading-screen/loading-screen.tsx
@@ -1,31 +1,29 @@
-import React, {Fragment} from "react";
+import * as React from "react";
 
-const ErrorScreen = () => {
+const LoadingScreen = () => {
   return (
-    <Fragment>
-      <div className="user-page">
-        <header className="page-header user-page__head">
+    <React.Fragment>
+      <section className="movie-card">
+
+        <h1 className="visually-hidden">WTW</h1>
+
+        <header className="page-header movie-card__head">
           <div className="logo">
-            <a href="main.html" className="logo__link">
+            <a className="logo__link">
               <span className="logo__letter logo__letter--1">W</span>
               <span className="logo__letter logo__letter--2">T</span>
               <span className="logo__letter logo__letter--3">W</span>
             </a>
           </div>
-
-          <h1 className="page-title user-page__title">Loading Error</h1>
         </header>
 
-        <div className="sign-in user-page__content">
-          <div className="sign-in__message">
-            <p>Sorry, something went wrong.</p>
-            <p> Please try again later.</p>
-          </div>
-        </div>
+        <h3 style={{textAlign: `center`, marginTop: `180px`, marginBottom: `160px`}}>Loading...</h3>
+      </section>
 
+      <div className="page-content">
         <footer className="page-footer">
           <div className="logo">
-            <a href="main.html" className="logo__link logo__link--light">
+            <a className="logo__link logo__link--light">
               <span className="logo__letter logo__letter--1">W</span>
               <span className="logo__letter logo__letter--2">T</span>
               <span className="logo__letter logo__letter--3">W</span>
@@ -37,9 +35,8 @@ const ErrorScreen = () => {
           </div>
         </footer>
       </div>
-
-    </Fragment>
+    </React.Fragment>
   );
 };
 
-export default ErrorScreen;
+export default LoadingScreen;
