+++ src/components/movie-card/movie-card.tsx
@@ -1,7 +1,7 @@
-import React from "react";
+import * as React from "react";
 import {Link} from "react-router-dom";
 import PropTypes from "prop-types";
-import {AppRoute} from "../../const.js";
+import {AppRoute} from "../../const";
 import {movieShape} from "../shapes";
