+++ src/components/no-auth-route/no-auth-route.tsx
@@ -1,11 +1,11 @@
-import React from "react";
+import * as React from "react";
 import {Route, Redirect} from "react-router-dom";
 import {connect} from "react-redux";
 import PropTypes from "prop-types";
 
-import {AppRoute} from "../../const.js";
-import {AuthorizationStatus} from "../../reducer/user/user.js";
-import {getAuthorizationStatus} from "../../reducer/user/selectors.js";
+import {AppRoute} from "../../const";
+import {AuthorizationStatus} from "../../reducer/user/user";
+import {getAuthorizationStatus} from "../../reducer/user/selectors";
 
 const NoAuthRoute = (props) => {
   const {authorizationStatus, path, render} = props;
