+++ src/components/show-more-button/show-more-button.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import * as React from "react";
 import PropTypes from "prop-types";
 
 const ShowMoreButton = (props) => {
