+++ src/hocs/with-active-item/with-active-item.tsx
@@ -1,8 +1,8 @@
-import React, {PureComponent} from "react";
+import * as React from "react";
 import PropTypes from "prop-types";
 
 const withActiveItem = (Component) => {
-  class WithActiveItem extends PureComponent {
+  class WithActiveItem extends React.PureComponent {
     constructor(props) {
       super(props);
       this.state = {
