+++ src/reducer/app-state/selectors.js
@@ -1,4 +1,4 @@
-import NameSpace from "../name-space.js";
+import NameSpace from "../name-space";
 
 const NAME_SPACE = NameSpace.APP_STATE;
