+++ src/reducer/reducer.js
@@ -1,8 +1,8 @@
 import {combineReducers} from "redux";
-import {reducer as data} from "./data/data.js";
-import {reducer as appState} from "./app-state/app-state.js";
-import {reducer as user} from "./user/user.js";
-import NameSpace from "./name-space.js";
+import {reducer as data} from "./data/data";
+import {reducer as appState} from "./app-state/app-state";
+import {reducer as user} from "./user/user";
+import NameSpace from "./name-space";
 
 export default combineReducers({
   [NameSpace.DATA]: data,
