+++ src/reducer/user/selectors.js
@@ -1,5 +1,5 @@
-import NameSpace from "../name-space.js";
 import {createSelector} from "reselect";
+import NameSpace from "../name-space";
 
 const NAME_SPACE = NameSpace.USER;
